Simplify the upgrade modal
