+++ webapp/src/components/backstage/upgrade_modal.tsx
@@ -3,21 +3,18 @@ import {useSelector} from 'react-redux';
 
 import styled from 'styled-components';
 
-import moment from 'moment';
-
-import {getCurrentUser} from 'mattermost-redux/selectors/entities/users';
-import General from 'mattermost-redux/constants/general';
-
 import GenericModal from 'src/components/widgets/generic_modal';
-import UpgradeIllustrationSvg from 'src/components/assets/upgrade_illustration_svg';
-import Spinner from 'src/components/assets/icons/spinner';
-import UpgradeSuccessIllustrationSvg from 'src/components/assets/upgrade_success_illustration_svg';
 import {requestTrialLicense, postMessageToAdmins} from 'src/client';
-import StartTrialNotice from 'src/components/backstage/start_trial_notice';
+import UpgradeModalFooter from 'src/components/backstage/upgrade_modal_footer';
 
-import {getAdminAnalytics} from 'src/selectors';
+import {isCurrentUserAdmin, getAdminAnalytics} from 'src/selectors';
 
 import {AdminNotificationType} from 'src/constants';
+
+import {ModalActionState, getUpgradeModalButtons, getUpgradeModalCopy} from 'src/components/backstage/upgrade_modal_data';
+
+import UpgradeModalIllustrationWrapper from 'src/components/backstage/upgrade_modal_illustration';
+import UpgradeModalHeader from 'src/components/backstage/upgrade_modal_header';
 
 interface Props {
     messageType: AdminNotificationType;
@@ -25,151 +22,13 @@ interface Props {
     onHide: () => void;
 }
 
-const isSystemAdmin = (roles: string): boolean => {
-    const rolesArray = roles.split(' ');
-    return rolesArray.includes(General.SYSTEM_ADMIN_ROLE);
-};
-
-type HandlerType = undefined | (() => (Promise<void> | void));
-
-interface ModalContents {
-    illustration: React.ReactNode;
-    titleText: string;
-    helpText: React.ReactNode;
-    confirmButtonText : React.ReactNode;
-    cancelButtonText : React.ReactNode;
-    handleConfirm : HandlerType;
-    handleCancel : HandlerType;
-}
-
-enum ModalActionState {
-    Uninitialized,
-    Loading,
-    Error,
-    Success,
-}
-
-const getModalData = (onHide: () => void, requestLicense: HandlerType, notifyAdmins: HandlerType, isAdmin: boolean, state: ModalActionState, messageType: AdminNotificationType) : ModalContents => {
-    let titleText = '';
-    let helpText = '';
-
-    switch (messageType) {
-    case AdminNotificationType.PLAYBOOK:
-        titleText = 'Playbook limit reached';
-        helpText = 'Every incident is different. With multiple playbooks each incident\'s workflow can be refined over time to improve time to resolution.';
-        break;
-    case AdminNotificationType.VIEW_TIMELINE:
-    case AdminNotificationType.MESSAGE_TO_TIMELINE:
-        titleText = 'Add more to your timeline';
-        helpText = 'Add important messages from the incident channel to the timeline and improve context in your retrospective.';
-        break;
-    }
-
-    const CommonModal = {
-        illustration: <UpgradeIllustrationSvg/>,
-        titleText,
-        helpText,
-        confirmButtonText: 'Start trial',
-        cancelButtonText: 'Not right now',
-        handleConfirm: requestLicense,
-        handleCancel: onHide,
-    };
-
-    const AdminStartModal = {...CommonModal};
-
-    const AdminLoadingModal = Object.assign({...AdminStartModal}, {
-        cancelButtonText: <Spinner/>,
-        // eslint-disable-next-line no-undefined
-        handleConfirm: undefined,
-        handleCancel: () => { /*do nothing*/ },
-    });
-
-    const AdminErrorModal = Object.assign({...AdminStartModal}, {
-
-    });
-
-    const expiryDate = moment().add('days', 30).format('MMMM D, YYYY');
-
-    const AdminSuccessModal = Object.assign({...AdminStartModal}, {
-        illustration: <UpgradeSuccessIllustrationSvg/>,
-        titleText: 'Your 30-day trial has started',
-        helpText: (
-            <span>
-                {`Your Enterprise E10 license expires on ${expiryDate}. You can purchase a license at any time through the `}
-                <a
-                    href='https://customers.mattermost.com/signup'
-                    target={'_blank'}
-                    rel='noreferrer'
-                >{'Customer Portal'}</a>
-                {' to avoid any disruption.'}
-            </span>
-        ),
-        confirmButtonText: 'Done',
-        handleConfirm: onHide,
-        // eslint-disable-next-line no-undefined
-        handleCancel: undefined,
-    });
-
-    const UserStartModal = Object.assign({...CommonModal}, {
-        helpText: helpText + ' Notify your System Admin to upgrade.',
-        confirmButtonText: 'Notify Administrator',
-        handleConfirm: notifyAdmins,
-    });
-
-    const UserLoadingModal = Object.assign({...UserStartModal}, {
-        cancelButtonText: <Spinner/>,
-        // eslint-disable-next-line no-undefined
-        handleConfirm: undefined,
-        handleCancel: () => { /*do nothing*/ },
-    });
-
-    const UserSuccessModal = Object.assign({...UserStartModal}, {
-        illustration: <UpgradeSuccessIllustrationSvg/>,
-        titleText: 'Thank you!',
-        helpText: 'Your System Admin has been notified',
-        confirmButtonText: 'Done',
-        handleConfirm: onHide,
-        // eslint-disable-next-line no-undefined
-        handleCancel: undefined,
-    });
-
-    if (isAdmin) {
-        switch (state) {
-        case ModalActionState.Uninitialized:
-            return AdminStartModal;
-        case ModalActionState.Loading:
-            return AdminLoadingModal;
-        case ModalActionState.Error:
-            return AdminErrorModal;
-        case ModalActionState.Success:
-            return AdminSuccessModal;
-        default:
-            return AdminStartModal;
-        }
-    } else {
-        switch (state) {
-        case ModalActionState.Uninitialized:
-            return UserStartModal;
-        case ModalActionState.Loading:
-            return UserLoadingModal;
-        case ModalActionState.Error:
-            return UserStartModal;
-        case ModalActionState.Success:
-            return UserSuccessModal;
-        default:
-            return UserStartModal;
-        }
-    }
-};
-
 const UpgradeModal: FC<Props> = (props: Props) => {
-    const currentUser = useSelector(getCurrentUser);
-    const isCurrentUserAdmin = isSystemAdmin(currentUser.roles);
+    const isAdmin = useSelector(isCurrentUserAdmin);
+
+    const [actionState, setActionState] = useState(ModalActionState.Uninitialized);
 
     const analytics = useSelector(getAdminAnalytics);
     const serverTotalUsers = analytics?.TOTAL_USERS || 0;
-
-    const [actionState, setActionState] = useState(ModalActionState.Uninitialized);
 
     const requestLicense = async () => {
         if (actionState === ModalActionState.Loading) {
@@ -198,7 +57,8 @@ const UpgradeModal: FC<Props> = (props: Props) => {
         setActionState(ModalActionState.Success);
     };
 
-    const modalData = getModalData(props.onHide, requestLicense, notifyAdmins, isCurrentUserAdmin, actionState, props.messageType);
+    const copy = getUpgradeModalCopy(isAdmin, actionState, props.messageType);
+    const buttons = getUpgradeModalButtons(isAdmin, actionState, requestLicense, notifyAdmins, props.onHide);
 
     return (
         <SizedGenericModal
@@ -206,110 +66,34 @@ const UpgradeModal: FC<Props> = (props: Props) => {
             show={props.show}
             modalHeaderText={''}
             onHide={props.onHide}
-            confirmButtonText={modalData.confirmButtonText}
-            cancelButtonText={modalData.cancelButtonText}
-            handleCancel={modalData.handleCancel}
-            handleConfirm={modalData.handleConfirm}
+            confirmButtonText={buttons.confirmButtonText}
+            cancelButtonText={buttons.cancelButtonText}
+            handleCancel={buttons.handleCancel}
+            handleConfirm={buttons.handleConfirm}
             autoCloseOnConfirmButton={false}
             footer={(
-                <Footer
+                <UpgradeModalFooter
                     actionState={actionState}
-                    isCurrentUserAdmin={isCurrentUserAdmin}
+                    isCurrentUserAdmin={isAdmin}
                 />
             )}
         >
             <Content>
-                <IllustrationWrapper>
-                    {modalData.illustration}
-                </IllustrationWrapper>
-                <Header>
-                    <Title>{modalData.titleText}</Title>
-                    <HelpText>{modalData.helpText}</HelpText>
-                </Header>
+                <UpgradeModalIllustrationWrapper
+                    state={actionState}
+                />
+                <UpgradeModalHeader
+                    titleText={copy.titleText}
+                    helpText={copy.helpText}
+                />
             </Content>
         </SizedGenericModal>
-    );
-};
-
-const FooterContainer = styled.div`
-    min-height: 32px;
-    width: 362px;
-    height: 32px;
-
-    font-size: 11px;
-    line-height: 16px;
-
-    display: flex;
-    align-items: center;
-    text-align: center;
-
-    color: rgba(var(--center-channel-color, 0.56));
-
-    margin-top: 18px;
-`;
-
-interface FooterProps {
-    actionState: ModalActionState;
-    isCurrentUserAdmin: boolean;
-}
-
-const Footer : FC<FooterProps> = (props: FooterProps) => {
-    if (!props.isCurrentUserAdmin) {
-        return null;
-    }
-
-    if (props.actionState !== ModalActionState.Uninitialized) {
-        return null;
-    }
-
-    return (
-        <FooterContainer>
-            <StartTrialNotice/>
-        </FooterContainer>
     );
 };
 
 const Content = styled.div`
     display: flex;
     flex-direction: column;
-`;
-
-const CenteredRow = styled.div`
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-`;
-
-const IllustrationWrapper = styled(CenteredRow)`
-    height: 156px;
-`;
-
-const Header = styled.div`
-    display: flex;
-    flex-direction: column;
-    margin-top: 20px;
-`;
-
-const Title = styled(CenteredRow)`
-    display: grid;
-    align-content: center;
-    height: 32px;
-    margin-bottom: 8px;
-
-    font-weight: 600;
-    font-size: 24px;
-    color: rgba(var(--center-channel-color-rgb), 1);
-`;
-
-const HelpText = styled(CenteredRow)`
-    display: grid;
-    align-content: center;
-    text-align: center;
-    width: 448px;
-
-    font-weight: 400;
-    font-size: 12px;
-    color: var(--center-channel-color);
 `;
 
 const SizedGenericModal = styled(GenericModal)`
